# Optimize zap.Any to use less memory on stack (inline version)