@cdvr1993 2023-07-28
Even though this doesn't use unsafe, I feel it is still harder to maintain than #1303. Here you are dropping all the typed constructors. Then if in the future anything changes on how we created Fields, we would need to update in 2 places: constructors and Any().