@abhinav 2023-07-28
> Even though this doesn't use unsafe, I feel it is still harder to maintain than #1303. Here you are dropping all the typed constructors. Then if in the future anything changes on how we created Fields, we would need to update in 2 places: constructors and Any().

Agreed! I don't think we should merge this as-is.
If I have time this weekend, I may explore auto-generating the unrolled version by having it inspect the current implementation.