@rabbbit 2023-07-29
I pushed a slightly smaller version of this to https://github.com/uber-go/zap/compare/pawel/any-int5 - it reduced the LOC changed around a half. It does seem slightly easier to parse than ^.